fix: move RequiresBilling from config.Library to serviceconfig.API
