+++ internal/librarian/java/repometadata.go
@@ -108,8 +108,13 @@ func deriveRepoMetadata(cfg *config.Config, library *config.Library, sourceDir s
 		LibraryType:          repometadata.GAPICAutoLibraryType,
 	}
 
-	if library.RequiresBilling != nil {
-		metadata.RequiresBilling = *library.RequiresBilling
+	apiCfg, err := serviceconfig.Find(sourceDir, library.APIs[0].Path, config.LanguageJava)
+	if err != nil {
+		return nil, fmt.Errorf("failed to find api config: %w", err)
+	}
+
+	if apiCfg.RequiresBilling != nil {
+		metadata.RequiresBilling = *apiCfg.RequiresBilling
 	} else if library.Java != nil {
 		metadata.RequiresBilling = !library.Java.BillingNotRequired
 	} else {
