+++ internal/librarian/library_test.go
@@ -1090,7 +1090,6 @@ func TestResolvePreview(t *testing.T) {
 				SkipGenerate:        false,
 				SkipRelease:         false,
 				SpecificationFormat: "protobuf",
-				RequiresBilling:     func() *bool { b := true; return &b }(),
 				Go: &config.GoModule{
 					ModulePathVersion: "v1",
 				},
@@ -1105,7 +1104,6 @@ func TestResolvePreview(t *testing.T) {
 					SkipGenerate:        true,
 					SkipRelease:         true,
 					SpecificationFormat: "discovery",
-					RequiresBilling:     func() *bool { b := false; return &b }(),
 					Go: &config.GoModule{
 						NestedModule: "v2",
 					},
@@ -1122,7 +1120,6 @@ func TestResolvePreview(t *testing.T) {
 				SkipGenerate:        true,
 				SkipRelease:         true,
 				SpecificationFormat: "discovery",
-				RequiresBilling:     func() *bool { b := false; return &b }(),
 				Go: &config.GoModule{
 					ModulePathVersion: "v1",
 					NestedModule:      "v2",
